prueba_deploy_25

+++ pyramids_yapo/app/utils/query.py
@@ -23,7 +23,7 @@ class RePyramidsYapoQuery:
             a.ad_id_nk,
             a.email,
             a.price::bigint,
-            a.uf_price,
+            a.uf_price::int,
             case
             when a.uf_price >= 1 AND a.uf_price < 3000 then '0-3000UF'
             when a.uf_price >= 3000 AND a.uf_price < 5000 then '3000-5000UF'
